requirements.txt
pip==21.3.1
tensorflow-gpu==2.7.0
tensorflow-hub==0.12.0
matplotlib==3.5.1
